ALFREDAPI-463: Migrate some integration tests to unit tests
